Fix modis_frp GEE export: cast bands to float (UInt32 fire_count vs Float32 FRP)
GEE rejects mixed band types on image export; the smoke test caught modis_frp failing. .toFloat() unifies frp_mean/frp_max/fire_count.

+++ Desktop/ISROPROJECT/src/isro_aqi/ingestion/modis_fire.py
@@ -37,11 +37,13 @@ def active_fire_frp(cfg: Config, start: str, end: str) -> "ee.Image":
         .filterBounds(region)
     )
     frp = coll.map(lambda i: i.divide(10.0))  # scale to MW
+    # .toFloat(): count() is UInt32 while the FRP bands are Float32, and GEE's
+    # image export requires all bands to share a data type -- cast to one.
     return ee.Image.cat(
         frp.mean().rename("frp_mean"),
         frp.max().rename("frp_max"),
         coll.count().rename("fire_count"),
-    )
+    ).toFloat()
 
 
 def burned_area(cfg: Config, start: str, end: str) -> "ee.Image":
